move logger to utils/logger
