+++ utils/index.ts
@@ -1,1 +1,2 @@
 export * from './utils';
+export * from './logger';
