+++ utils/logger.ts
@@ -1,0 +1,17 @@
+import pino from "pino";
+
+const transport = pino.transport({
+  target: 'pino-pretty',
+});
+
+export const logger = pino(
+  {
+    level: 'info',
+    redact: ['poolKeys'],
+    serializers: {
+      error: pino.stdSerializers.err,
+    },
+    base: undefined,
+  },
+  transport,
+);
